Btrfs: u64 cleanups

+++ fs/btrfs/kerncompat.h
@@ -16,7 +16,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #define BUG() abort()
 
 typedef unsigned int u32;
-typedef unsigned long u64;
+typedef unsigned long long u64;
 typedef unsigned char u8;
 typedef unsigned short u16;
